add filter by type
