+++ src/components/data-api/magicCards-api.js
@@ -3,7 +3,7 @@ import { createAsyncThunk } from '@reduxjs/toolkit';
 const createCardsList = (obj) => {
   const result = [];
 
-  obj.cards.slice(0, 20).forEach((item) => {
+  obj.cards.slice(0, 50).forEach((item) => {
     if (item.imageUrl) {
       result.push({
         convertedManaCost: item.cmc,
